analysis/pyproject.toml
[project]
name = "analysis"
version = "0.1.0"
requires-python = "~=3.12.0"
dependencies = [
    "fire>=0.7.1",
    "jinja2>=3.1.6",
    "logdir>=0.13.0",
    "loguru>=0.7.3",
    "numpy>=2.3.5",
    "pandas>=2.3.3",
    "pingouin>=0.5.5",
    "python-slugify>=8.0.4",
    "ribs[all]==0.9.0",
    "ruamel-yaml>=0.18.17",
    "seaborn>=0.13.2",
    "statsmodels>=0.14.6",
    "tqdm>=4.67.1",
]

[dependency-groups]
dev = [
    "pyinstrument>=5.1.1",
    "pylint>=4.0.4",
    "pytest>=9.0.2",
    "pytest-cov>=7.0.0",
    "ruff>=0.14.9",
]
